+++ sol/src/Event/KeyboardEvent.h
@@ -10,15 +10,15 @@ namespace sol
 struct KeyboardEvent : public Event
 {
 protected:
-	uint8_t key_code;
+	uint16_t key_code;
 
-	KeyboardEvent(uint8_t key_code)
+	KeyboardEvent(uint16_t key_code)
 	    : key_code(key_code)
 	{
 	}
 
 public:
-	uint8_t get_key_code() const { return key_code; }
+	uint16_t get_key_code() const { return key_code; }
 
 	EVENT_CLASS_CATEGORY(INPUT | KEYBOARD)
 };
@@ -26,16 +26,16 @@ public:
 struct KeyPressedEvent : public KeyboardEvent
 {
 private:
-	uint8_t repeat_count = 0;
+	uint16_t repeat_count = 0;
 
 public:
-	KeyPressedEvent(uint8_t key_code, uint8_t repeat_count = 0)
+	KeyPressedEvent(uint16_t key_code, uint16_t repeat_count = 0)
 	    : KeyboardEvent(key_code)
 	    , repeat_count(repeat_count)
 	{
 	}
 
-	uint8_t get_repeat_count() const { return repeat_count; }
+	uint16_t get_repeat_count() const { return repeat_count; }
 
 	virtual std::string to_string() const override
 	{
@@ -51,7 +51,7 @@ public:
 struct KeyReleasedEvent : public KeyboardEvent
 {
 public:
-	KeyReleasedEvent(uint8_t key_code)
+	KeyReleasedEvent(uint16_t key_code)
 	    : KeyboardEvent(key_code)
 	{
 	}
